Memoizing selectCollection and collectionUrlParam installed lodash library memoize
 import memoize from 'lodash.memoize';

wrap our selectCollection function with memoize:
  export const selectCollection = memoize(collectionUrlParam => (
  createSelector(
    [selectCollections],
    collections =>
      collections.find(collection => collection.id === COLLECTIION_ID_MAP[collectionUrlParam])
      )
    ));

lodash.memoize does the same idea of memoization as reselect does
for our selectors, except this time we're memoizing the return of
our function which returns our selector

By wrapping this function is memoize, we're saying that whenever
this function gets called and receives collectionUrlParam, I want
to memoize the return of this function (in this case we return a
selector). If this function gets called again with the same
collectionUrlParam, don't rerun this function because we'll return
the same value as last time, which we've memoized so just return
the selector that's been stored.

Changes to be committed:
	modified:   package.json
	modified:   src/redux/shop/shop.selectors.js